Crate canyon_primitives

Low-level types used throughout the Canyon code.

Modules

report
App-specific crypto used for reporting equivocation/misbehavior in BABE and GRANDPA. Any rewards for misbehavior reporting will be paid out to this account.

Constants

PERMASTORE_INHERENT_IDENTIFIER
Unique identifier for the permastore Inherent

POA_INHERENT_IDENTIFIER
Unique identifier for the poa Inherent

Type Definitions

AccountId
Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

AccountIndex
The type for looking up accounts. We don’t expect more than 4 billion of them.

Balance
Balance of an account.

Block
Block type.

BlockId
Block ID.

BlockNumber
An index to a block.

DataIndex
The type for indexing the perma storage. The upper limit is around 16 EiB.

Depth
Number of the attempts of PoA generation.

DigestItem
Digest item type.

ExtrinsicIndex
Type used for extrinsic.

Hash
A hash of some data used by the chain.

Header
Header type.

Index
Index of a transaction in the chain.

Moment
Type used for expressing timestamp.

Signature
Alias to 512-bit hash when used in the context of a transaction signature on the chain.

Timestamp
A timestamp: milliseconds since the unix epoch. u64 is enough to represent a duration of half a billion years, when the time scale is milliseconds.